requirements.txt
streamlit
pandas
qrcode
streamlit-authenticator
python-dotenv
openpyxl
boto3
sqlalchemy
psycopg2-binary==2.9.9
tabulate
opencv-python
streamlit-webrtc 
av
streamlit-qrcode-scanner
streamlit-js-eval 
opencv-python-headless
streamlit-aggrid
